+++ gui/components/LoadingSymbol.js
