+++ gui/components/NavTitle.js
