+++ gui/components/Quotes.js
