+++ gui/components/grid/GameBoard.js
@@ -2,10 +2,10 @@ const m = require("mithril");
 const classnames = require("classnames");
 
 
-import HexGrid from "./HexGrid.js";
-import Hud from "./GameHud.js";
-import Unit from "../models/Unit.js";
-import Tile from "../models/Tile.js";
+import HexGrid from "../HexGrid.js";
+import Hud from "../hud/GameHud.js";
+import Unit from "../../models/Unit.js";
+import Tile from "../../models/Tile.js";
 
 
 const GameBoard = (initialVnode) => {
