+++ gui/components/grid/HexTile.js
@@ -1,7 +1,7 @@
 const m = require("mithril");
 const classNames = require("classnames");
 
-import PlayableUnit from "./PlayableUnit.js";
+import PlayableUnit from "../unit/PlayableUnit.js";
 import Unit from "../../models/Unit.js";
 import scenarioDefinitionStore from "../../stores/ScenarioDefinition.store.js";
