+++ gui/pages/scenario/tile/BattleByTileId.js
@@ -1,8 +1,8 @@
 const m = require('mithril');
 
 import Page from '../../../models/Page.js';
-import BattleResultCalc from "../../../components/BattleResultCalc.js";
-import TitleBar from "../../../components/templates/TitleBar.js";
+import BattleResultCalc from "../../../components/battle/BattleResultCalc.js";
+import TitleBar from "../../../components/TitleBar.js";
 import Body from "../../../components/Body.js";
 import Background from "../../../components/Background.js";
 import ScenarioDefinitionStore from "../../../stores/ScenarioDefinition.store.js";
