+++ gui/pages/scenario/tile/MapByTileId.js
@@ -1,7 +1,7 @@
 const m = require('mithril');
 
 import Body from "../../../components/Body.js";
-import Button from "../../../components/templates/Button.js";
+import Button from "../../../components/Button.js";
 import Background from "../../../components/Background.js";
 import Page from '../../../models/Page.js';
 import ScenarioDefinitionStore from "../../../stores/ScenarioDefinition.store.js";
