+++ gui/pages/settings/ScenarioList.js
@@ -3,9 +3,9 @@ const classnames = require('classnames');
 
 import Page from '../../models/Page.js';
 import Body from "../../components/Body.js";
-import Button from "../../components/templates/Button.js";
+import Button from "../../components/Button.js";
 import scenarioManifestStore from "../../stores/ScenarioManifest.store.js";
-import TitleBar from "../../components/templates/TitleBar.js";
+import TitleBar from "../../components/TitleBar.js";
 
 
 export const page = new Page('/settings/scenarios', (initialVnode) => {
